Kirillb/more fixes

+++ src/main/java/dts/controllers/AdminController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import boundaries.OperationBoundary;
 import boundaries.UserBoundary;
+import dts.data.UserRole;
 import dts.logic.ItemService;
 import dts.logic.OperationService;
 import dts.logic.UsersService;
@@ -25,46 +26,79 @@ public class AdminController {
 	
 	@Autowired
 	public AdminController(ItemService itemService, UsersService usersService, OperationService operationService) {
-
 		this.itemService = itemService;
 		this.usersService = usersService;
 		this.operationService = operationService;
 	}
 
-	@RequestMapping(method = RequestMethod.DELETE, path = "/dts/admin/users/{adminSpace}/{adminEmail}")
-	public void deleteAllUsers(@PathVariable("adminSpace") String adminSpace,
+	@RequestMapping(
+			method = RequestMethod.DELETE, 
+			path = "/dts/admin/users/{adminSpace}/{adminEmail}")
+	public void deleteAllUsers(
+			@PathVariable("adminSpace") String adminSpace,
 			@PathVariable("adminEmail") String adminEmail) {
-		this.itemService.deleteAll(adminSpace, adminEmail);
+		if (this.isAdmin(adminSpace, adminEmail)) 
+			this.usersService.deleteAllUsers(adminSpace, adminEmail);
 	}
 
-	@RequestMapping(method = RequestMethod.DELETE, path = "/dts/admin/items/{adminSpace}/{adminEmail}")
-	public void deleteAllItems(@PathVariable("adminSpace") String adminSpace,
+	@RequestMapping(
+			method = RequestMethod.DELETE, 
+			path = "/dts/admin/items/{adminSpace}/{adminEmail}")
+	public void deleteAllItems(
+			@PathVariable("adminSpace") String adminSpace,
 			@PathVariable("adminEmail") String adminEmail) {
-		this.itemService.deleteAll(adminSpace, adminEmail);
+		if (this.isAdmin(adminSpace, adminEmail)) 
+			this.itemService.deleteAll(adminSpace, adminEmail);
 	}
 
-	@RequestMapping(method = RequestMethod.DELETE, path = "/dts/admin/operations/{adminSpace}/{adminEmail}")
-	public void deleteAllOperations(@PathVariable("adminSpace") String adminSpace,
+	@RequestMapping(
+			method = RequestMethod.DELETE, 
+			path = "/dts/admin/operations/{adminSpace}/{adminEmail}")
+	public void deleteAllOperations(
+			@PathVariable("adminSpace") String adminSpace,
 			@PathVariable("adminEmail") String adminEmail) {
-		this.operationService.deleteAllActions(adminSpace, adminEmail);
+		if (this.isAdmin(adminSpace, adminEmail)) 
+			this.operationService.deleteAllActions(adminSpace, adminEmail);
 	}
 
-	@RequestMapping(method = RequestMethod.GET, path = "/dts/admin/users/{adminSpace}/{adminEmail}", produces = MediaType.APPLICATION_JSON_VALUE)
-	public UserBoundary[] exportAllUsers(@PathVariable("adminSpace") String adminSpace,
+	@RequestMapping(
+			method = RequestMethod.GET, 
+			path = "/dts/admin/users/{adminSpace}/{adminEmail}", 
+			produces = MediaType.APPLICATION_JSON_VALUE)
+	public UserBoundary[] exportAllUsers(
+			@PathVariable("adminSpace") String adminSpace,
 			@PathVariable("adminEmail") String adminEmail) {
-		List<UserBoundary> listUser = this.usersService.getAllUsers(adminSpace, adminEmail);
-		UserBoundary[] array = new UserBoundary[listUser.size()];
-		listUser.toArray(array);
+		UserBoundary[] array = {};
+		if (this.isAdmin(adminSpace, adminEmail)) {
+			List<UserBoundary> listUser = this.usersService.getAllUsers(adminSpace, adminEmail);
+			array = new UserBoundary[listUser.size()];
+			listUser.toArray(array);
+		}
 		return array;
 	}
 
-	@RequestMapping(method = RequestMethod.GET, path = "/dts/admin/operations/{adminSpace}/{adminEmail}", produces = MediaType.APPLICATION_JSON_VALUE)
-	public OperationBoundary[] exportAllOperations(@PathVariable("adminSpace") String adminSpace,
+	@RequestMapping(
+			method = RequestMethod.GET, 
+			path = "/dts/admin/operations/{adminSpace}/{adminEmail}", 
+			produces = MediaType.APPLICATION_JSON_VALUE)
+	public OperationBoundary[] exportAllOperations(
+			@PathVariable("adminSpace") String adminSpace,
 			@PathVariable("adminEmail") String adminEmail) {
-		List<OperationBoundary> listBoundary = this.operationService.getAllOperations(adminSpace, adminEmail);
-		OperationBoundary[] array = new OperationBoundary[listBoundary.size()];
-		listBoundary.toArray(array);
+		OperationBoundary[] array = {};
+		if (this.isAdmin(adminSpace, adminEmail)) {
+			List<OperationBoundary> listBoundary = this.operationService.getAllOperations(adminSpace, adminEmail);
+			array = new OperationBoundary[listBoundary.size()];
+			listBoundary.toArray(array);
+		}
 		return array;
 	}
 
+	private boolean isAdmin(String adminSpace, String adminEmail) {
+		if (this.usersService.login(adminSpace, adminEmail).getRole().equals(UserRole.ADMIN)) 
+			return true;
+		else {
+			System.err.println("User" + adminEmail + "is not an Admin user");
+			return false;
+		}
+	}
 }
